+++ ultrack-prototype/frontend/task.ts
@@ -1,0 +1,6 @@
+export type Task = {
+  index: number;
+  question: string;
+  answers: string[];
+  answerIndex?: number;
+};
